Fixes #1936 - Preproc constructor cast fix

+++ core/src/processing/core/PApplet.java
@@ -9635,7 +9635,7 @@ public class PApplet extends Applet
   }
   */
 
-  static final public float[] parseByte(byte what[]) {
+  static final public float[] parseFloat(byte what[]) {
     float floaties[] = new float[what.length];
     for (int i = 0; i < what.length; i++) {
       floaties[i] = what[i];
